consolidate schemas
