+++ packages/api/router/orgPhoto.ts
@@ -1,6 +1,6 @@
+import { CreateOrgPhotoSchema, UpdateOrgPhotoSchema } from 'schemas/create/orgPhoto'
 import { defineRouter, handleError, permissionedProcedure } from '~api/lib'
 import { CreateAuditLog } from '~api/schemas/create/auditLog'
-import { CreateOrgPhotoSchema, UpdateOrgPhotoSchema } from '~api/schemas/orgPhoto'
 
 export const orgPhotoRouter = defineRouter({
 	create: permissionedProcedure
